[DEV][add] Change stack/queue sizes to 8192*8
Also fix merge conflict on Makefile

Note:
Everything thus far is a mix of static and dynamic allocations.
It wouldn't cost much performance-wise to dynamically allocate
queues and stacks at creation and a reallocate later if they get
filled.

A strategy could consist of growing at 80% and shrinking at 25%,
in the background.